2025-5a: Solved this.
I wanted to a big set of integers for this, but even the first range
exhausted Python's memory, and even killed my terminal emulator!

I also explored expanding overlapping ranges and coalescing adjacent
ranges, but didn't implement these.

In the end, just creating the 185 ranges and checking ingredients in
all of those ranges until a hit or we checked all 185 worked quickly.